[TNL-5632] Don't request responses/comments for non-AJAX requests.
When a user loads a single thread in discussions by directly navigating
to the discussions URL, or refreshing the page, it triggers the LMS to
load the given comment thread behind-the-scenes.

This request is sent without any pagination parameters -- because those
are generated behind-the-scenes in the JavaScript code that actually
drives the discussions pages.  This leads to loading entire threads,
some of which are veryyyy large, and essentially never using them to
render the actual page: an AJAX request happens no matter what when the
page loads, renders a properly-paginated result set.

This change simply checks to see if we're in an AJAX request and
requests the thread to come with responses accordingly.  If we're not in
an AJAX request, we don't want the responses, otherwise, we do.