[TNL-5632] Remove get_threads calls when rendering a single thread.
Before, the view was getting a list of threads for the course and
shipping them back to the caller.  Since rendering a single thread will
start a user off with a list of topics, and not threads, we were getting
a list of threads and then not using them at all.

We've removed this call and switched to a static list that only contains
the current thread being displayed.  Unfortunately, we need to know the
thread details before the UI can then make the call to fetch responses,
so we're unable to remove the "get this thread" call when the in
non-AJAX mode.