[TNL-5632] Add a cleaner way to use the request cache, and cache course structures.
We created a new decorator, request_cached, that automatically forms a
cache key based on the method name and positional arguments, and
automatically applies caching logic to the return value of the decorated
function.  It's not applicable for every case  -- we still have plenty
of places where the boilerplate exists -- but many of those use cases
can take advantage of this decorator.

Using this decorator, we're also now request caching the "cached" course
structure that gets polled when loading a single thread.