[TNL-5632] Tweaks to fix tests that dependned us single_thread getting a thread list.
After removing the API call for grabbing the list of threads in the
single_thread view, we broke some tests which expected that call to be
the last call made by invoking that view.  We also broke some tests that
looked at query counts by changing the single_thread view as well as
adding request caching to different bits.

With these items tweaked, we're cleaning up a lot of test failures.